+++ libreplan-business/src/main/java/org/libreplan/business/email/entities/EmailTemplate.java
@@ -27,9 +27,7 @@ import org.libreplan.business.settings.entities.Language;
  * This class is intended to work as a Hibernate component.
  * It represents the E-mail template to be modified by admin and send to user.
  *
- * Created by
- * @author Vova Perebykivskiy <vova@libreplan-enterprise.com>
- * on 29.09.2015.
+ * @author Created by Vova Perebykivskyi <vova@libreplan-enterprise.com> on 29.09.2015.
  */
 public class EmailTemplate extends BaseEntity {
 
@@ -56,14 +54,14 @@ public class EmailTemplate extends BaseEntity {
     }
 
     public String getContent() {
-        return content;
+        return content != null ? content : "";
     }
     public void setContent(String content) {
         this.content = content;
     }
 
     public String getSubject() {
-        return subject;
+        return subject != null ? subject : "";
     }
     public void setSubject(String subject) {
         this.subject = subject;