+++ libreplan-webapp/src/main/java/org/libreplan/importers/notifications/realization/SendEmailOnResourceRemovedFromTask.java
@@ -36,12 +36,10 @@ import java.util.List;
 
 /**
  * Sends E-mail to users with data that storing in notification_queue table
- * and that are treat to TEMPLATE_RESOUCE_REMOVED_FROM_TASK
+ * and that are treat to {@link EmailTemplateEnum#TEMPLATE_RESOURCE_REMOVED_FROM_TASK}
  * Data will be send if resource has been removed from task (in resource allocation)
  *
- * Created by
- * @author Vova Perebykivskiy <vova@libreplan-enterprise.com>
- * on 20.01.2016.
+ * @author Created by Vova Perebykivskiy <vova@libreplan-enterprise.com> on 20.01.2016.
  */
 
 @Component
@@ -61,14 +59,11 @@ public class SendEmailOnResourceRemovedFromTask implements IEmailNotificationJob
     public void sendEmail() {
         // At this time all data have gathered, if it exists of course
 
-        if ( Configuration.isEmailSendingEnabled() ){
+        if ( Configuration.isEmailSendingEnabled() ) {
+            if ( emailConnectionValidator.isConnectionActivated() && emailConnectionValidator.validConnection()  ) {
 
-            if ( emailConnectionValidator.isConnectionActivated() )
-
-                if ( emailConnectionValidator.validConnection()  ){
-
-                List<EmailNotification> notifications = emailNotificationModel
-                        .getAllByType(EmailTemplateEnum.TEMPLATE_RESOURCE_REMOVED_FROM_TASK);
+                List<EmailNotification> notifications =
+                        emailNotificationModel.getAllByType(EmailTemplateEnum.TEMPLATE_RESOURCE_REMOVED_FROM_TASK);
 
                 for (int i = 0; i < notifications.size(); i++)
                     if ( composeMessageForUser(notifications.get(i)) )